bootloader: Add 'die_err' function
This commit saved 4 bytes of space, and made the code more
extensible (as if it could be extended, anyway).

Furthermore, the "Not found" error could be turned into an error
code as well, but I rather keep it this way, since the error can be more
user-related than system-related (like the BIOS ones).

* Add 'error_codes.asm' file.